Refactoring: add directories for modules
